Saving more computational time after crossvalidation.
After we cross validate, the new images coming from the NEB calculation
are computed with the DFT calculator and AMP calculator. Therefore, in
order to append them to the training set, I have created a new
trajectory file called images_from_neb.traj that will be opened with the
Trajectory class, and appended to  the training set without the need of
calling set_calculators method.