A new restarting case has been added in this commit.
In the case that a ML-NEB calculation was succesful but
a crossvalidation failes, the `restart` keyword will take care of this.
It works in the following way: it reads all file in the directory. Then,
it checks the latest neb_%s.traj on disk. It slices the last files, and
perform cross_validation, and then continues.